requirements.txt
reportportal-client
robotframework>=4
robotframework-debuglibrary
robotframework-seleniumlibrary
robotframework-jupyterlibrary>=0.3.1
ipython
openshift==0.12.1
pre-commit
pytest
pytest-logger
pyyaml
pygments
requests
Jinja2
flake8
mypy
kubernetes
validators
